refactor(codehealth): diagrams package to green
